Adds default sidekiq config; starts sidekiq if APP_ROLE == 'sidekiq' (#220)
Starts sidekiq based on the APP_ROLE variable.

+++ start.sh
@@ -1,12 +1,17 @@
-set -e 
+set -e
 
 if [ ${RAILS_ENV:-development} != "production" ]; then
   bundle check || bundle
 fi
 
-echo "starting rails"
-rm -f tmp/pids/server.pid
-bundle exec rails db:create
-bundle exec rails db:migrate
-bundle exec rails solr:init
-bundle exec rails s -b '0.0.0.0'
+if [ ${APP_ROLE:-app} == "sidekiq" ]; then
+    echo "starting sidekiq"
+    /app/bin/sidekiq
+else
+    echo "starting rails"
+    rm -f tmp/pids/server.pid
+    bundle exec rails db:create
+    bundle exec rails db:migrate
+    bundle exec rails solr:init
+    bundle exec rails s -b '0.0.0.0'
+fi
